рефакторит структуру
