+++ js/components/abstract-view.js
@@ -1,4 +1,8 @@
-export default `
+import {createElement} from '../utils';
+
+class AbstractView {
+  constructor() {
+    this.footerTemplate = `\
 <footer class="footer">
   <a href="https://htmlacademy.ru" class="social-link social-link--academy">HTML Academy</a>
   <span class="footer__made-in">Сделано в <a href="https://htmlacademy.ru" class="footer__link">HTML Academy</a> &copy; 2016</span>
@@ -8,5 +12,28 @@ export default `
     <a href="https://www.facebook.com/htmlacademy" class="social-link  social-link--fb">Фэйсбук</a>
     <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
   </div>
-</footer>
-`;
+</footer>`;
+  }
+
+  getTemplate() {
+    throw new Error(`You have to define template for view`);
+  }
+
+  render() {
+    return createElement(this.getTemplate());
+  }
+
+  bind() {
+
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = this.render();
+      this.bind();
+    }
+    return this._element;
+  }
+}
+
+export default AbstractView;
