+++ js/components/greeting/greeting-view.js
@@ -21,7 +21,7 @@ class GreetingView extends AbstractView {
   </div>
   <div class="greeting__continue"><span><img src="img/arrow_right.svg" width="64" height="64" alt="Next"></span></div>
 </div>
-${this._footerTemplate}\
+${this.footerTemplate}\
   `;
   }
