+++ js/components/intro/intro.js
@@ -1,0 +1,17 @@
+import IntroView from './intro-view';
+import changeView from '../../router/change-view';
+
+export default () => {
+  /**
+   * Переключает на следующий экран
+   * @param {MouseEvent} evt
+   */
+  const onNextButtonClick = (evt) => {
+    evt.preventDefault();
+    changeView(`greeting`);
+  };
+
+  const intro = new IntroView(onNextButtonClick);
+
+  return intro;
+};
