+++ js/components/stats/stats.js
@@ -1,8 +1,7 @@
 import StatsView from './stats-view';
-import changeView from '../router/change-view';
 import onBackButtonClicked from '../handlers/back-button-handler';
-import getScore from '../models__old/get-score';
-import {state} from '../data/state';
+import getScore from '../../data/get-score';
+import {state} from '../../data/state';
 
 
 export default () => {
