+++ js/main.js
@@ -1,5 +1,3 @@
 import changeView from './router/change-view';
-// import InitialState from './models/initial-state';
 
-// const state = new InitialState();
 changeView(`intro`);
