+++ js/utils.js
@@ -3,7 +3,7 @@
  * @param {string} template
  * @return {Element}
  */
-export default (template) => {
+export const createElement = (template) => {
   const element = document.createElement(`div`);
   element.innerHTML = template;
   return element;
